.travis.yml:  use latest stable Go (1.13.7 at this time)
This is possible because the GopherJS compiler supports being
built with newer versions of Go as of gopherjs/gopherjs#966.

As of Go 1.12, go tool vet is no longer supported.¹
Start using the go vet command instead.

Use RCS format for gofmt diff. It results in cleaner,
easier to read gofmt diffs.

¹ https://golang.org/doc/go1.12#vet